+++ app/src/main/java/com/bgcoding/notes/app/feature_note/presentation/notes/NotesScreen.kt
@@ -7,34 +7,31 @@ import androidx.compose.animation.fadeOut
 import androidx.compose.animation.slideInVertically
 import androidx.compose.animation.slideOutVertically
 import androidx.compose.foundation.background
-import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.safeContentPadding
 import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.Notes
 import androidx.compose.material.icons.automirrored.filled.Sort
 import androidx.compose.material.icons.filled.Add
-import androidx.compose.material.icons.filled.CameraAlt
 import androidx.compose.material.icons.filled.Delete
+import androidx.compose.material.icons.filled.DeleteForever
 import androidx.compose.material.icons.filled.Menu
-import androidx.compose.material.icons.filled.Notes
 import androidx.compose.material.icons.filled.Search
 import androidx.compose.material.icons.filled.Settings
-import androidx.compose.material.icons.filled.Sort
-import androidx.compose.material3.Button
+import androidx.compose.material3.AlertDialogDefaults
+import androidx.compose.material3.BasicAlertDialog
 import androidx.compose.material3.DrawerValue
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.FloatingActionButton
@@ -50,21 +47,24 @@ import androidx.compose.material3.SnackbarDuration
 import androidx.compose.material3.SnackbarHost
 import androidx.compose.material3.SnackbarHostState
 import androidx.compose.material3.SnackbarResult
+import androidx.compose.material3.Surface
 import androidx.compose.material3.SwipeToDismissBox
 import androidx.compose.material3.SwipeToDismissBoxValue
 import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
 import androidx.compose.material3.TopAppBar
 import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.material3.rememberDrawerState
 import androidx.compose.material3.rememberSwipeToDismissBoxState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.DisposableEffect
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.rememberUpdatedState
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
@@ -79,12 +79,91 @@ import kotlinx.coroutines.launch
 @Composable
 fun NotesScreen(
     navController: NavController,
-    viewModel: NotesViewModel = hiltViewModel()
+    viewModel: NotesViewModel = hiltViewModel(),
 ) {
+
+    // get showDeleted from navController argument
+    val showDeleted = navController.currentBackStackEntry?.arguments?.getBoolean("showDeleted") ?: false
+
+    LaunchedEffect(showDeleted) {
+        viewModel.onEvent(NotesEvent.SetShowDeleted(showDeleted))
+        Log.d("NotesScreen", "in LaunchedEffect showDeleted: $showDeleted")
+    }
+
     val state = viewModel.state.value
     val snackbarHostState = remember { SnackbarHostState() }
     val drawerState = rememberDrawerState(DrawerValue.Closed)
     val scope = rememberCoroutineScope()    // scoped to the composable
+
+    val showDialog = remember { mutableStateOf(false) }
+
+    if (showDialog.value) {
+        BasicAlertDialog(
+            onDismissRequest = { showDialog.value = false }
+        ) {
+            Surface(
+                modifier = Modifier
+                    .padding(16.dp),
+                shape = MaterialTheme.shapes.large,
+                tonalElevation = AlertDialogDefaults.TonalElevation
+            ) {
+                Column(
+                    modifier = Modifier.wrapContentSize()
+                ) {
+                    Text(
+                        text = "Delete Notes",
+                        style = MaterialTheme.typography.titleLarge,
+                        modifier = Modifier.padding(16.dp).align(Alignment.CenterHorizontally),
+                        textAlign = TextAlign.Center
+                    )
+                    Text(
+                        text = "Are you sure you want to permanently delete all notes?",
+                        style = MaterialTheme.typography.bodyMedium,
+                        modifier = Modifier.padding(horizontal = 16.dp),
+                        textAlign = TextAlign.Center
+                    )
+                    Spacer(modifier = Modifier.height(4.dp))
+                    Row(
+                        modifier = Modifier
+                            .fillMaxWidth()
+                            .padding(
+                                start = 16.dp,
+                                end = 16.dp,
+                                bottom = 16.dp
+                            ),
+                        horizontalArrangement = Arrangement.SpaceAround
+                    ) {
+                        TextButton(
+                            onClick = {
+                                showDialog.value = false
+                            }
+                        ) {
+                            Text(
+                                "Cancel",
+                                style = MaterialTheme.typography.bodyMedium,
+                                color = MaterialTheme.colorScheme.error,
+                                textAlign = TextAlign.Center
+                            )
+                        }
+                        Spacer(modifier = Modifier.width(8.dp))
+                        TextButton(
+                            onClick = {
+                                showDialog.value = false
+                                viewModel.onEvent(NotesEvent.DeleteAllNotesPermanently)
+                            }
+                        ) {
+                            Text(
+                                "Delete",
+                                style = MaterialTheme.typography.bodyMedium,
+                                color = MaterialTheme.colorScheme.error,
+                                textAlign = TextAlign.Center
+                            )
+                        }
+                    }
+                }
+            }
+        }
+    }
 
     // Log recompositions
     val currentState = rememberUpdatedState(state)
@@ -92,6 +171,10 @@ fun NotesScreen(
         Log.d("NotesScreen", "Recomposed with state: $currentState")
         onDispose { }
     }
+
+
+
+    Log.d("NotesScreen", "showDeleted: $showDeleted")
 
     ModalNavigationDrawer(
         drawerState = drawerState,
@@ -122,8 +205,10 @@ fun NotesScreen(
                     label = {
                         Text(text = "Notes", style = MaterialTheme.typography.bodyMedium, fontWeight = FontWeight.Bold)
                     },
-                    selected = true,
-                    onClick = { /*TODO*/ }
+                    selected = !showDeleted,
+                    onClick = {
+                        navController.navigate(Screen.NotesScreen.route + "?showDeleted=false")
+                    }
                 )
                 NavigationDrawerItem(
                     icon = {
@@ -135,8 +220,10 @@ fun NotesScreen(
                     label = {
                         Text(text = "Bin", style = MaterialTheme.typography.bodyMedium, fontWeight = FontWeight.Bold)
                     },
-                    selected = false,
-                    onClick = { /*TODO*/ }
+                    selected = showDeleted,
+                    onClick = {
+                        navController.navigate(Screen.NotesScreen.route + "?showDeleted=true")
+                    }
                 )
                 NavigationDrawerItem(
                     icon = {
@@ -204,12 +291,21 @@ fun NotesScreen(
                 )
             },
             floatingActionButton = {
-                FloatingActionButton(onClick = {
-                    navController.navigate(Screen.AddEditNoteScreen.route)
+                FloatingActionButton(onClick =
+                {
+                    if (!showDeleted) {
+                        navController.navigate(Screen.AddEditNoteScreen.route)
+                    } else {
+                        showDialog.value = true
+                    }
                 },
                     containerColor = MaterialTheme.colorScheme.primaryContainer,
                 ) {
-                    Icon(imageVector = Icons.Filled.Add, contentDescription = "Add note")
+                    if (!showDeleted)
+                        Icon(imageVector = Icons.Filled.Add, contentDescription = "Add note")
+                    else {
+                        Icon(imageVector = Icons.Filled.DeleteForever, contentDescription = "Delete Notes")
+                    }
                 }
             },
             snackbarHost = { SnackbarHost(snackbarHostState) },
@@ -238,13 +334,13 @@ fun NotesScreen(
                     .background(MaterialTheme.colorScheme.surface)
                 ) {
                     items(state.notes, key = { it.id!! }) { note ->
-                        /*DisposableEffect(Unit) {
+                        DisposableEffect(Unit) {
                             Log.d("LazyColumn", "LazyColumn recomposed")
                             onDispose {}
-                        }*/
+                        }
                         val dismissState = rememberSwipeToDismissBoxState(
                             confirmValueChange = {
-                                if (it == SwipeToDismissBoxValue.EndToStart) {
+                                if (!showDeleted && it == SwipeToDismissBoxValue.EndToStart) {
                                     viewModel.onEvent(NotesEvent.DeleteNote(note))
                                     scope.launch {
                                         val result = snackbarHostState
@@ -256,12 +352,22 @@ fun NotesScreen(
                                             )
                                         when (result) {
                                             SnackbarResult.ActionPerformed -> {
-                                                viewModel.onEvent(NotesEvent.RestoreNote)
+                                                viewModel.onEvent(NotesEvent.RestorePreviouslyDeletedNote)
                                             }
                                             SnackbarResult.Dismissed -> {
                                                 // do nothing
                                             }
                                         }
+                                    }
+                                    true
+
+                                } else if (showDeleted && it == SwipeToDismissBoxValue.EndToStart) {
+                                    viewModel.onEvent(NotesEvent.RestoreDeletedNote(note))
+                                    scope.launch {
+                                        snackbarHostState.showSnackbar(
+                                            message = "Note restored",
+                                            duration = SnackbarDuration.Short
+                                        )
                                     }
                                     true
                                 } else {
@@ -273,19 +379,37 @@ fun NotesScreen(
                         SwipeToDismissBox(
                             state = dismissState,
                             backgroundContent = {
-                                Box(
-                                    modifier = Modifier
-                                        .fillMaxSize()
-                                        .background(MaterialTheme.colorScheme.error)
-                                        .padding(horizontal = 20.dp),
-                                    contentAlignment = Alignment.CenterEnd
-                                ) {
-                                    Text(
-                                        text = "Delete",
-                                        style = MaterialTheme.typography.bodyMedium,
-                                        color = MaterialTheme.colorScheme.onError
-                                    )
+                                if (!showDeleted) {
+                                    Box(
+                                        modifier = Modifier
+                                            .fillMaxSize()
+                                            .background(MaterialTheme.colorScheme.error)
+                                            .padding(horizontal = 20.dp),
+                                        contentAlignment = Alignment.CenterEnd
+                                    ) {
+                                        Text(
+                                            text = "Delete",
+                                            style = MaterialTheme.typography.bodyMedium,
+                                            color = MaterialTheme.colorScheme.onError
+                                        )
+                                    }
+                                } else {
+                                    Box(
+                                        modifier = Modifier
+                                            .fillMaxSize()
+                                            .background(MaterialTheme.colorScheme.secondaryContainer)
+                                            .padding(horizontal = 20.dp),
+                                        contentAlignment = Alignment.CenterEnd
+                                    ) {
+                                        Text(
+                                            text = "Restore",
+                                            style = MaterialTheme.typography.bodyMedium,
+                                            color = MaterialTheme.colorScheme.onSecondaryContainer
+                                        )
+                                    }
                                 }
+
+
                             }
                         ) {
                             NoteItem(
@@ -297,7 +421,6 @@ fun NotesScreen(
                                         navController.navigate(
                                             Screen.AddEditNoteScreen.route +
                                                     "?noteId=${note.id}"
-//                                        "?noteId=${note.id}&noteColor=${note.color}"
                                         )
                                     }
                             )
